+++ core/processors/client_processor_test.go
